# api/openapi.yml
openapi: 3.1.0
info:
  title: MOBIUS.NVR HTTP API
  version: "0.1.0"
  summary: |
    Contract between the MOBIUS.NVR Flask backend and its native clients
    (`~/0_MOBIUS.NVR.CLIENT.DESKTOP/`, `~/0_MOBIUS.NVR.CLIENT.MOBILE/`).
  description: |
    Phase 0 deliverable for the native-client plan
    (`docs/plans/native_desktop_and_mobile_client_alongside_browser_frontend_evaluating_pyside6_flutter_gtk_with_gstreamer_video_open_source_only.md`).

    This document is the canonical API contract. Both client repos copy it in
    at build time (via `tools/sync-api-contract.sh`) and codegen models
    (`datamodel-code-generator` for Python, `openapi-generator` for Dart).

    ## Scope

    - All blueprint-registered Flask routes the browser frontend consumes
      (`routes/*.py`, `services/cert_routes.py`).
    - Excludes `services/external_api_routes.py` (third-party Bearer-token
      surface, separately documented at `/api/external/docs`).

    ## Phase 0 vs Phase 1+

    Path / method / auth / params are pinned in Phase 0. Response payload
    schemas are intentionally permissive (`additionalProperties: true` on
    most domain objects) — they tighten in Phase 1 as each endpoint is
    first consumed by the desktop client. Exact field lists at that point
    are pulled from `routes/*.py` source rather than guessed.

    ## Auth model

    Two cookies are involved:

    - `session` — Flask-Login session cookie (HttpOnly, SameSite=Lax).
...
      `POST /logout`.
...

...
    captcha/2fa).

...
servers:
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask-Login session cookie. Issued by `POST /login`.
    DeviceCookie:
      type: apiKey
      in: cookie
      name: device_token
      description: Long-lived per-device UUID cookie. Set on first login,
        max_age=365 days. Used for device trust + heartbeat.
  parameters:
...
    CameraId:
      name: camera_id
      in: path
      required: true
      description: Same as `camera_serial`. Some legacy routes use
        `camera_id` in the path; semantics are identical.
      schema:
        type: string
        minLength: 4
...
  responses:
...
    NotFound:
      description: Resource (camera, user, recording, evidence event, ...)
        does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: |
            Human-readable error message. The backend is consistent in
            returning `{"error": "..."}` — but a few legacy routes also use
            `{"success": false, "error": "..."}`; clients should accept either.
        success:
          type: boolean
          description: Present on legacy routes only.
    SuccessEnvelope:
      type: object
      properties:
        status:
          type: string
          examples: ["saved", "ok", "started", "stopped"]
        success:
          type: boolean
      additionalProperties: true

    # ---- Domain objects ---------------------------------------------------
    Camera:
      type: object
      description: Camera config row from the `cameras` DB table. Per
        CLAUDE.md, the DB is the runtime source of truth; `cameras.json`
        is a seed file.
      properties:
        serial:        { type: string, description: Primary key. }
        name:          { type: string, description: Display name. }
        brand:         { type: string, examples: ["reolink", "eufy", "amcrest", "sv3c", "unifi", "neolink"] }
        model:         { type: string }
        ip:            { type: string }
        rtsp_input:    { type: string, description: Source RTSP URL (creds redacted server-side). }
        stream_type:   { $ref: "#/components/schemas/StreamType" }
        streaming_hub: { $ref: "#/components/schemas/StreamingHub" }
        ptz:           { type: boolean }
        audio:         { type: boolean }
        audio_input_supported: { type: boolean, description: Set by Phase 1 ffprobe survey of evidence pipeline. }
        hidden:        { type: boolean }
        ui_health_monitor: { type: boolean }
        capabilities:  { type: array, items: { type: string } }
        video_fit_mode: { $ref: "#/components/schemas/VideoFitMode" }
      additionalProperties: true
    StreamType:
      type: string
      enum:
        - WEBRTC
        - HLS
        - LL_HLS
        - MJPEG
        - NEOLINK
        - NEOLINK_LL_HLS
      description: Per-camera default stream protocol. User may override
        via `user_camera_preferences`.
    StreamingHub:
      type: string
      enum: [mediamtx, go2rtc]
      description: Which relay hub serves this camera. `mediamtx` is the
        default; `go2rtc` covers Baichuan (Reolink E1, etc.) and the
        in-progress consolidation target.
    VideoFitMode:
      type: string
      nullable: true
      enum: [cover, contain, fill, null]
...
security:
  - SessionCookie: []
    DeviceCookie: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /streams:
    get:
      tags: [Pages]
      summary: Multi-stream viewing page (HTML).
      description: Auto-redirects mobile/tablet UAs to `/light` unless `?full=1`.
      parameters:
        - { name: full, in: query, schema: { type: string }, description: "Pass `1` to bypass mobile redirect." }
      responses:
        "200": { description: HTML. }
        "302": { description: Redirected to /light for mobile UAs. }
...
  /api/cameras/{camera_id}:
    parameters: [{ $ref: "#/components/parameters/CameraId" }]
    get:
      tags: [Cameras]
      summary: Get a single camera's metadata.
      responses:
        "200":
          description: Camera record.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Camera" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/settings/trusted-network:
    get:
      tags: [Settings]
      summary: Trusted-network status (auto-login when on LAN, etc.).
      responses:
        "200":
          description: Trusted-network blob.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    put:
      tags: [Settings]
      summary: Update trusted-network setting (admin only).
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessEnvelope" }
  /api/settings/streaming-hubs:
    get:
      tags: [Settings, Stream Control]
      summary: Per-vendor + global streaming-hub assignments.
      responses:
        "200":
          description: Hub assignment blob.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    put:
      tags: [Settings, Stream Control]
      summary: Update streaming-hub assignments (admin only).
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessEnvelope" }
...
  /api/settings/advanced/{key}:
    parameters:
      - { name: key, in: path, required: true, schema: { type: string } }
    patch:
      tags: [Settings]
      summary: Patch one advanced setting (admin only).
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessEnvelope" }

  # ============================== PTZ ======================================
...